HS-77 프리 함수 리팩토링 (#44)
* HS-77 REFACT: env.c -> env_utils.c

* HS-77 REFACT: safe_free 를 free_utils  폴더로 이동

* HS-77 RENAME: safe_free -> free_utils

* HS-77 WIP: free_all function

* HS-77 FIX: 빈값이 입력될 때 할당도 안하기 때문에 프리도 하면 안 됨

* HS-77 REFACT: 관련있는 라인끼리 모으기

* HS-77 FEAT: 함수간 종속성 제거

* HS-77 FIX: all free, multi process

* HS-77 WIP: free...
