+++ src/init.c
@@ -17,6 +17,7 @@ void    init_minishell(t_minishell *minishell)
 			exit(EXIT_SUCCESS);
 		add_history(input);
 		replaced_input = replace_whole_input_dollar(input, minishell);
+		printf("%s%s%s\n", GRN,replaced_input,COLOR_RESET);
 		if (replaced_input)
 		{
 			if (*replaced_input)
@@ -24,36 +25,13 @@ void    init_minishell(t_minishell *minishell)
 				splitted_input = command_split(replaced_input);
 				curr_token = make_token_list(splitted_input);
 				set_process_list(&(minishell->ps_list), curr_token);
-				minishell->ps_list->cmd_line = curr_token;
+				// minishell->ps_list->cmd_line = curr_token; //  필요없는 코드로 판단. 필요없는게 맞다면 주석 지워주길 바람.(작성하신 분이). seongyle
 				// executor(minishell); wait 추가되야  함.
 			}
 		}
 		else
 			break ;
-		free(replaced_input);
-		replaced_input = NULL;
-		i = -1;
-		while (splitted_input[++i])
-			free(splitted_input[i]);
-		free(splitted_input);
-		splitted_input = NULL;
-		t_token *temp;
-		while (curr_token)
-		{
-			temp = curr_token;
-			free(temp);
-			curr_token = curr_token->next;
-		}
-		t_process	*curr_process;
-		t_process	*temp_process;
-		curr_process = minishell->ps_list;
-		while (curr_process)
-		{
-			temp_process = curr_process;
-			curr_process = curr_process->next;
-			free(temp_process);
-			temp_process = NULL;
-		}
-		system("leaks minishell");
+		free_all(minishell, replaced_input, splitted_input);
+		// system("leaks minishell");
 	}
 }
