+++ src/parser/make_arr_to_list.c
@@ -27,7 +27,7 @@ static void	insert_first_into_list(t_token **lst, char *str)
 
 	temp = safe_malloc(sizeof(t_token));
 	temp->type = 0;
-	temp->value = str;
+	temp->value = ft_strdup(str);
 	temp->next = NULL;
 	temp->prev = NULL;
 	check_and_set_type(temp);
@@ -40,7 +40,7 @@ static void	insert_str_into_list_back(t_token **lst, char *str)
 
 	temp = safe_malloc(sizeof(t_token));
 	temp->type = 0;
-	temp->value = str;
+	temp->value = ft_strdup(str);
 	temp->next = NULL;
 	temp->prev = *lst;
 	check_and_set_type(temp);
